added hex view of directories and sections

+++ PEExplorerV2/HexControl/InMemoryBuffer.h
@@ -7,10 +7,10 @@ public:
 	InMemoryBuffer();
 
 	// Inherited via IBufferManager
-	virtual uint32_t GetData(int64_t offset, uint8_t * buffer, uint32_t count) override;
-	virtual void Insert(int64_t offset, const uint8_t * data, uint32_t count) override;
-	virtual void Delete(int64_t offset, uint32_t count) override;
-	virtual void SetData(int64_t offset, const uint8_t * data, uint32_t count) override;
+	virtual size_t GetData(int64_t offset, uint8_t * buffer, size_t count) override;
+	virtual void Insert(int64_t offset, const uint8_t * data, size_t count) override;
+	virtual void Delete(int64_t offset, size_t count) override;
+	virtual void SetData(int64_t offset, const uint8_t * data, size_t count) override;
 	virtual int64_t GetSize() const override;
 	virtual uint8_t* GetRawData(int64_t offset) override;
